FeedRepository: Add the logic and tests for fetching the feed data
Fetch the feed data by using the FeedApiService.
After the deta is fetched, we map it to FeedItems and update the value of the repository LiveData.

The network error thrown while fetching the data implementation includes notifying the above layer
by updating the value of a LiveData and logging the thrown error message.

Add dependency injection to FeedRepository and FeedViewModel.
Change implementation to use RxJava objects.

strings.xml: Add repository network error text.